Revert "add beta chrome extension + update next"
