--- apps/extension/public/manifest.json
@@ -1,25 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/chrome-manifest",
-  "manifest_version": 3,
-  "name": "Card Confluence Extension",
-  "version": "0.0.0",
-  "description": "Replaces deckbox search with Scryfall autocomplete on Moxfield.",
-  "permissions": ["storage"],
-  //   "host_permissions": [
-  // "https://scryfall.com/*",
-  //   "https://api.scryfall.com/*",
-  //   "https://scryfall.com/docs/tagger-tags",
-  // "https://tagger.scryfall.com/*"
-  //   ],
-  "background": {
-    "service_worker": "background.js",
-    "type": "module"
-  },
-  "content_scripts": [
-    {
-      "matches": ["https://moxfield.com/decks/*"],
-      "js": ["content.js"],
-      "run_at": "document_idle"
-    }
-  ]
-}
