+++ apps/web/package.json
@@ -10,45 +10,45 @@
   },
   "dependencies": {
     "@codemirror/autocomplete": "6.18.6",
-    "@codemirror/commands": "^6.10.0",
-    "@codemirror/language": "^6.11.3",
-    "@codemirror/view": "^6.39.2",
-    "@google/genai": "^1.32.0",
-    "@radix-ui/react-accordion": "^1.2.12",
-    "@radix-ui/react-dialog": "^1.1.15",
-    "@radix-ui/react-dropdown-menu": "^2.1.16",
-    "@radix-ui/react-label": "^2.1.8",
-    "@radix-ui/react-progress": "^1.1.8",
-    "@radix-ui/react-slider": "^1.3.6",
-    "@radix-ui/react-slot": "^1.2.4",
-    "@radix-ui/react-switch": "^1.2.6",
-    "@radix-ui/react-tooltip": "^1.2.8",
-    "@repo/scryfall-search": "workspace:*",
-    "@uiw/react-codemirror": "^4.25.3",
+    "@codemirror/commands": "^6.8.1",
+    "@codemirror/language": "^6.11.0",
+    "@codemirror/view": "^6.36.8",
+    "@google/genai": "^1.4.0",
+    "@radix-ui/react-accordion": "^1.2.11",
+    "@radix-ui/react-dialog": "^1.1.14",
+    "@radix-ui/react-dropdown-menu": "^2.1.14",
+    "@radix-ui/react-label": "^2.1.7",
+    "@radix-ui/react-progress": "^1.1.7",
+    "@radix-ui/react-slider": "^1.3.5",
+    "@radix-ui/react-slot": "^1.2.2",
+    "@radix-ui/react-switch": "^1.2.5",
+    "@radix-ui/react-tooltip": "^1.2.6",
+    "@uiw/react-codemirror": "^4.23.12",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "codemirror-lang-scrycards": "workspace:*",
     "lucide-react": "^0.511.0",
-    "next": "16.0.8",
+    "next": "15.3.2",
     "next-themes": "^0.4.6",
-    "react": "^19.2.1",
-    "react-dom": "^19.2.1",
+    "node-html-parser": "^7.0.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "react-scrycards": "^1.7.0",
-    "tailwind-merge": "^3.4.0"
+    "tailwind-merge": "^3.3.0"
   },
   "devDependencies": {
-    "@eslint/eslintrc": "^3.3.3",
+    "@eslint/eslintrc": "^3",
     "@repo/eslint-config": "workspace:*",
     "@repo/typescript-config": "workspace:*",
     "@scryfall/api-types": "1.0.0-alpha.4",
-    "@tailwindcss/postcss": "^4.1.17",
-    "@types/node": "^22.19.2",
-    "@types/react": "^19.2.7",
-    "@types/react-dom": "^19.2.3",
-    "eslint": "^9.39.1",
+    "@tailwindcss/postcss": "^4",
+    "@types/node": "^22",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
+    "eslint": "^9",
     "eslint-config-next": "15.3.2",
-    "tailwindcss": "^4.1.17",
-    "tw-animate-css": "^1.4.0",
-    "typescript": "^5.9.3"
+    "tailwindcss": "^4",
+    "tw-animate-css": "^1.3.0",
+    "typescript": "^5"
   }
 }
