--- packages/scryfall-search/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@repo/scryfall-search",
-  "version": "0.0.0",
-  "main": "./src/index.ts",
-  "types": "./src/index.ts",
-  "sideEffects": false,
-  "dependencies": {
-    "@scryfall/api-types": "1.0.0-alpha.4",
-    "codemirror-lang-scrycards": "workspace:*"
-  },
-  "devDependencies": {
-    "@repo/typescript-config": "workspace:*",
-    "typescript": "^5"
-  }
-}
