--- packages/scryfall-search/src/index.ts
@@ -1,3 +1,0 @@
-export * from "./scryfall";
-export * from "./search";
-export type { ICatalog } from "codemirror-lang-scrycards";
