--- packages/scryfall-search/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "@repo/typescript-config/base.json",
-  "compilerOptions": {
-    "lib": ["ES2020", "DOM"],
-    "module": "nodenext",
-    "moduleResolution": "nodenext",
-    "target": "ES2020"
-  },
-  "include": ["src"],
-  "exclude": ["node_modules"]
-}
