+++ turbo.json
@@ -3,7 +3,6 @@
   "ui": "tui",
   "tasks": {
     "build": {
-      "env": ["GEMINI_API_KEY"],
       "dependsOn": ["^build"],
       "inputs": ["$TURBO_DEFAULT$", ".env*"],
       "outputs": [".next/**", "!.next/cache/**"]
